+++ src/routes/RequestsRoute.js
@@ -667,6 +667,7 @@ class RequestsRoute extends React.Component {
     const addressTypes = get(resources, 'addressTypes.records', []);
     const servicePoints = get(resources, 'servicePoints.records', []);
     const cancellationReasons = get(resources, 'cancellationReasons.records', []);
+    const requestCount = get(resources, 'records.other.totalRecords', 0);
     const InitialValues = dupRequest ||
       { requestType: 'Hold', fulfilmentPreference: 'Hold Shelf' };
 
@@ -691,6 +692,7 @@ class RequestsRoute extends React.Component {
         <Button
           buttonStyle="dropdownItem"
           id="exportToCsvPaneHeaderBtn"
+          disabled={!requestCount}
           onClick={() => {
             onToggle();
             if (!this.csvExportPending) {
@@ -699,17 +701,21 @@ class RequestsRoute extends React.Component {
             }
           }}
         >
-          <FormattedMessage id="stripes-components.exportToCsv" />
+          <FormattedMessage id="ui-requests.exportSearchResultsToCsv" />
         </Button>
         <Button
           buttonStyle="dropdownItem"
           id="exportExpiredHoldsToCsvPaneHeaderBtn"
+          disabled={!requestCount}
           onClick={() => {
             onToggle();
             this.exportExpiredHoldsToSCV();
           }}
         >
-          <FormattedMessage id="ui-requests.exportExpiredHoldsToCsv" />
+          <FormattedMessage
+            id="ui-requests.exportExpiredHoldShelfToCsv"
+            values={{ currentServicePoint: servicePointName }}
+          />
         </Button>
         <PrintButton
           buttonStyle="dropdownItem"